Refactor(eos_designs): Move eos_designs_facts to pyavd
