+++ python-avd/pyavd/_eos_designs/eos_designs_facts/__init__.py
@@ -5,10 +5,9 @@ from __future__ import annotations
 
 from functools import cached_property
 
-from ansible_collections.arista.avd.plugins.plugin_utils.avdfacts import AvdFacts
-from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
-
+from ...vendor.avdfacts import AvdFacts
+from ...vendor.errors import AristaAvdError
+from ...vendor.utils import get
 from .mlag import MlagMixin
 from .overlay import OverlayMixin
 from .short_esi import ShortEsiMixin
