+++ python-avd/pyavd/_eos_designs/eos_designs_facts/mlag.py
@@ -7,7 +7,7 @@ from functools import cached_property
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .eos_designs_facts import EosDesignsFacts
+    from . import EosDesignsFacts
 
 
 class MlagMixin:
