+++ python-avd/pyavd/_eos_designs/eos_designs_facts/wan.py
@@ -6,10 +6,10 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from ansible_collections.arista.avd.plugins.filter.natural_sort import natural_sort
+from ...j2filters.natural_sort import natural_sort
 
 if TYPE_CHECKING:
-    from .eos_designs_facts import EosDesignsFacts
+    from . import EosDesignsFacts
 
 
 class WanMixin:
